tuned up processing logic

+++ src/main/java/edu/mcw/rgd/pipelines/imexinteractions/Process.java
@@ -2,6 +2,7 @@ package edu.mcw.rgd.pipelines.imexinteractions;
 
 
 import edu.mcw.rgd.datamodel.Interaction;
+import edu.mcw.rgd.datamodel.InteractionAttribute;
 import edu.mcw.rgd.process.Utils;
 import psidev.psi.mi.tab.model.BinaryInteraction;
 
@@ -14,25 +15,34 @@ import java.util.*;
 public class Process {
     private int minFileSize = 27800;
     Dao dao = new Dao();
-    Date startDate = new Date();
     private String deleteThreshold;
 
-    public int processFile(String fileName) throws Exception {
+    public int processFile(String fileName, boolean processStaleData) throws Exception {
         long time0 = System.currentTimeMillis();
 
         int initialAttrCount = dao.getAttributeCount();
         List<Interaction> piList = loadProteinInteractions(fileName);
 
-        int insertedRecordCount = dao.insertOrUpdate(piList);
+        Collection<Interaction> piUnique = mergeDuplicateInteractions(piList);
 
-        int fileSize = dao.getFileSize(fileName);
-        if( fileSize>=getMinFileSize() ){
-             int deletedInteractionsCount = dao.deleteUnmodifiedInteractions(startDate);
-             System.out.println("Deleted Interaction Records Count: " + deletedInteractionsCount);
+        long time1 = System.currentTimeMillis();
+        int insertedRecordCount = dao.insertOrUpdate(piUnique);
+        System.out.println(" --- insertOrUpdate: elapsed "+Utils.formatElapsedTime(time1, System.currentTimeMillis()));
+
+        if( processStaleData ) {
+            // set the start date to yesterday: pipeline server and db server could be slightly out-of-sync,
+            // and we don't want to delete the just added interactions and attributes
+            Date startDate = Utils.addDaysToDate(new Date(), -1);
+
+            int fileSize = dao.getFileSize(fileName);
+            if (fileSize >= getMinFileSize()) {
+                int deletedInteractionsCount = dao.deleteUnmodifiedInteractions(startDate);
+                System.out.println("Deleted Interaction Records Count: " + deletedInteractionsCount);
+            }
+
+            int deletedInteractionAttributesCount = dao.deleteUnmodifiedInteractionAttributes(startDate, getDeleteThreshold(), initialAttrCount);
+            System.out.println("Deleted Interaction Attributes Count: " + deletedInteractionAttributesCount);
         }
-
-        int deletedInteractionAttributesCount = dao.deleteUnmodifiedInteractionAttributes(startDate, getDeleteThreshold(), initialAttrCount);
-        System.out.println("Deleted Interaction Attributes Count: " + deletedInteractionAttributesCount);
 
         System.out.println("===PROCESSING ELAPSED TIME: "+ Utils.formatElapsedTime(time0, System.currentTimeMillis()));
         return insertedRecordCount;
@@ -64,6 +74,38 @@ public class Process {
         System.out.println("Count of incoming unassigned pubmed values (duplicates possible): " + parser.countOfUnassignedPubmedValues);
 
         return piList1;
+    }
+
+    Collection<Interaction> mergeDuplicateInteractions( List<Interaction> interactions ) {
+        // unique interaction in rgd is identified by compound key: {rgdid1, rgdid2, interactiontype}
+        Map<String, Interaction> map = new HashMap<>(interactions.size());
+
+        for( Interaction i: interactions ) {
+            String compoundKey = i.getRgdId1()+"-"+i.getRgdId2()+"-"+i.getInteractionType();
+            Interaction uniqueI = map.get(compoundKey);
+            if( uniqueI==null ) {
+                map.put(compoundKey, i);
+            } else {
+                // only merge interaction attributes
+                uniqueI.getInteractionAttributes().addAll(i.getInteractionAttributes());
+            }
+        }
+        System.out.println("   ORIGINAL INCOMING INTERACTION COUNT: "+interactions.size());
+        System.out.println("   MERGED   INCOMING INTERACTION COUNT: "+map.size());
+
+        // merge attributes
+        int origAttrCount = 0;
+        int newAttrCount = 0;
+        for( Interaction i: map.values() ) {
+            origAttrCount += i.getInteractionAttributes().size();
+            Set<InteractionAttribute> attrs = new HashSet<>(i.getInteractionAttributes());
+            newAttrCount += attrs.size();
+            i.setInteractionAttributes(new ArrayList<>(attrs));
+        }
+
+        System.out.println("   ORIGINAL INCOMING ATTR COUNT: "+origAttrCount);
+        System.out.println("   MERGED   INCOMING ATTR COUNT: "+newAttrCount);
+        return map.values();
     }
 
     public int getMinFileSize() {
